# kratos_bubble/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Todo API
    version: 0.0.1
paths:
    /v1/todo:
        post:
            tags:
                - Todo
            operationId: Todo_CreateTodo
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateTodoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateTodoReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/todo/{id}:
        get:
            tags:
                - Todo
            operationId: Todo_GetTodo
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetTodoReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        put:
            tags:
                - Todo
            operationId: Todo_UpdateTodo
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateTodoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateTodoReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        delete:
            tags:
                - Todo
            operationId: Todo_DeleteTodo
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteTodoReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/todos:
        get:
            tags:
                - Todo
            operationId: Todo_ListTodo
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListTodoReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        CreateTodoReply:
            type: object
            properties:
                id:
                    type: string
                title:
                    type: string
                status:
                    type: boolean
        CreateTodoRequest:
            type: object
            properties:
                title:
                    type: string
        DeleteTodoReply:
            type: object
            properties: {}
        GetTodoReply:
            type: object
            properties:
                id:
                    type: string
                title:
                    type: string
                status:
                    type: boolean
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        ListTodoReply:
            type: object
            properties:
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/Todo'
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        Todo:
            type: object
            properties:
                id:
                    type: string
                title:
                    type: string
                status:
                    type: boolean
            description: 单独定义一个
        UpdateTodoReply:
            type: object
            properties: {}
        UpdateTodoRequest:
            type: object
            properties:
                id:
                    type: string
                title:
                    type: string
                status:
                    type: boolean
tags:
    - name: Todo